added copy paste support
added ability to copy and paste nodes both within a map and onto another map. If the page you are pasting the node in already has a node with a similar name, the node will be given a default name. Connections between copied nodes will be preserved, but connections to non-copied nodes will be removed when pasted